drm/i915: Handle framebuffer offsets[]
Take fb->offset[0] into account when calculating the linear and tile x/y
offsets.

For non-tiled surfaces fb->offset[0] is simply added to the linear
byte offset.

Treat fb->offsets[0] as a byte offset into the linearized view of the
surface. So we end up converting fb->offsets[0] into additional x and
y offsets.

In practice dealing with an extra X offset could get messy (eg. how
would it interact with the user specified stride?). So to keep things
simple, require offsets[] to be stride aligned. That way only the Y
offset and the height of the framebuffer as measured from the beginning
of the bo need to be considered.

v2: Put the computation into a new function, and do the conversion in
    a way that wouldn't result in out of bounds x coordinates.
    Handle SKL as well, and deal with the rotated GTT view.
    Require the offset to be stride aligned to keep things simple

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>