drm/i915: Implement atomic modesetting
Implement the mandatory hooks for the atomic modeset ioctl.

The code first makes a backup of the current state, then proceeds to
modify the state as properties are modified. After all the properties
have been handled the new state is checked, and if everything checks
out, the new state is commited to hardware. Finally we clean up any
temporary storage.

In theory everything is checked before the hardware state is clobbered,
so there should be no need for rollback. But as the current modesetting
code can still return errors from some of of the operation, the rollback
code is included.

v2:
* Use intel_best_encoder() directly
* Update new_crtc and new_encoder fields after atomic modeset
* Don't mark cursor as pinned, when we don't have a cursor bo
* Fix hiding the cursor with atomic ioctl
* Fix plane src rectangle dirty check
* Remove stale prototypes form atomic code
* Eliminate swap_old_new() hack in the atomic code
* Drop old_crtc and old_encoder pointers
* Add some TODO items to the atomic code
* Update connector DPMS state after an atomic modeset

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>