drm: Add a bunch of standard properties for CRTCs and planes
Several new properties for planes and CRTCs which will be used
by the atomic mode setting code.

For planes:
SRC_X/SRC_Y/SRC_W/SRC_H are the plane source coordinates. 16.16 fixed
point.
CRTC_X/CRTC_Y/CRTC_W/CRTC_H are the plane's destination coordinates.
CRTC_ID the currently assigned CRTC for this plane.
FB_ID the currently assigned FB for this plane

For CRTCs:
SRC_X/SRC_Y are the source coordinates.
FB_ID the currently assigned FB for this CRTC
MODE is the current display mode. It's a blob which contains
struct drm_mode_modeinfo.
CONNECTOR_IDS is a list of the current connectors. It's a blob
and contains an array of uint32_t connector IDs.
CURSOR_ID contains the GEM object ID of the cursor data.
CURSOR_X/CURSOR_Y are the cursor's destination coordinates.
CURSOR_W/CURSOR_H are the size of the cursor.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>