drm/i915: Reorder i915_gem_execbuffer_move_to_gpu() and i915_switch_context()
The FBC RT address is stored in the context, and thus needs to be
rewritten after a context switch before any batches are run. We emit
the LRI to update the FBC RT address when we call the ring ->flush
function to invalidate the caches.

When a context switch is being performed we currently do the
invalidate before the context switch, which means we're not updating
the FBC RT address correctly. Move the i915_gem_execbuffer_move_to_gpu()
call to happen after i915_switch_context() to make the LRI happen
after the MI_SET_CONTEXT.

v2: Rebased

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>