drm/i915: Add full pipe rotation
We can pretend that we can rotate the entire pipe by rotating all the
planes and adjusting their positions appropriately. Add a "crtc-rotation"
property on the crtc which will do this.

The main upshot of doing the full pipe rotation instead of rotating all
the planes individually is that the plane positions turn out correct
automagically. So userspace doesn't need to do anything except toggle
the property and continue as if nothing had changed.

The actual implementation is pretty much trivial thanks to drm_rect
and drm_rotation_chain() ;)

v2: Rename the property to "crtc-rotation""

Cc: Sagar Kamble <sagar.a.kamble@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>