dim: allow fixed sha1 for nightly rebuilding
We want to avoid the merge-window/early -rc craziness and so updated
drm-upstream for -nightly only manually. But that in turn caused
coordination headaches.

Add an optional sha1 to the tree specifier in nightly.conf to fix a
tree to a given revision. That way we can freeze drm-* to the last
version before the merge window and update it at our leisure.

With this fixed we can lift the ban and go back to auto-updating
drm-upstream.

v2: Pimp commit message and drop now outdated comment (Jani)