drm/i915: Reject fbc on g4x when sprites are enabled
WaFbcDisabledForOverlaySprite:ctg says we need to disallow fbc when
sprites are enabled.

We don't currently expose the sprites on g4x, but maybe we will later.
Add the appropriate check to fbc2_possible(). We will also need to
adjust the sprite code to turn off fbc prior to enabling the sprite,
but let's leave that until we actually get g4x sprite support.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>