dim: clean up dim checkpatch output
Use a local var for the commit for clarity. Print oneline commit summary
before the output. Use checkpatch.pl -q option to reduce output. Reduce
newlines printed for errors and warnings. Return 1 instead of exit 1 on
errors so we can continue as needed.

All of this is to better support other uses for dim_checkpatch in follow
up work.