drm/i915: Trunk gate display clocks on VLV when not needed
VLV Punit doesn't trunk gate the display clocks at CCK when it
power gates the disp2d well. The docs seem to be telling us to
manually trunk gate the clocks. We could do that for cdclk,
however CCK refuses to cooperate when it comes to the hrawclk.
Writes to the relevant register (0x6c) are simply ignored. But
we can trick Punit into doing the trunk gating by asking it to
clock gate the power well prior to power gating it. This results
in both cdclk and hrawclk getting trunk gated.

CHV Punit seems smarter and trunk gates the clocks when asked to
power gate the well.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>