drm/i915: Rework VLV DPIO power well handling
The hardware requirement on VLV is that the DPIO TX wells be brought up
before the CMN well. That means we can't really utilize the TX wells in
a dynamic fashion. Doing so would potentiall require toggling the CMN
well off and back on when bringing up more TX wells, and that could
wreak havoc on anyone else currently depending on the CMN well.

So let's stop pretending we could do this, and just collapse the TX
wells into a single entity with the CMN well. So from now on we bring
the entire PHY up/down as a single unit. This may actaully speed things
up a bit since we no longer bring each well up separately.

To avoid wasting too much power, we can however permanently power gate
any TX well that we know is never going to be needed. This would be the
case when one of the ports is not even present on the board, or if
there's an eDP panel attached that uses only one or two lanes.

During driver init we have to power up all the TX wells since we don't
yet know which ones we will need. During output setup, we can drop off
any useless TX well, and mark them as such so that they will stay off
permanently.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>