drm/i915: Power down the DSI PLL before reconfiguring it
The BIOS may leave the DSI PLL enabled in some wonky state where it just
refuses to lock. Simply disabling the PLL before reconfiguring it is not
enough to fix it, but power gating the PLL prior to reconfiguring does
work.

This happens on BYT FFRD8 when booting with HDMI connected so the DSI
display will not be lit up by the BIOS.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>