drm/i915: Always enable DPIO lane power overrides on CHV
Recent BIOS versions already enable the dynamic power down mode for the
PHY lanes. In order to power up the lanes we need to enable the
overrides in the PHY_CONTROL registers. This would properly done as
needed, but to keep things simple, just engage the overrides whenever
the PHY is powered on. This should make it behave as if the dynamic
power down mode was still disabled.

v2: Set the overrides already in chv_phy_control_init()
v3: Fix ~ fumble when ORing bits

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>