WIP: drm/i915: Type safe register read/write
Make I915_READ and I915_WRITE more type safe by wrapping the register
offset in a struct. This should eliminate most of the fumbles we've had
with misplaced parens.

This only takes care of normal mmio registers. We could extend the idea
to other register types and define each with its own struct. That way
you wouldn't be able to accidentally pass the wrong thing to a specific
register access function.

There are a few uglies left:
- switch statements don't like structs as case values, even if
  you case 'case DEADBEEF.reg:'. Fortunately we don't have too many
  of those so can maybe switch them to use some port enums or such,
  or if all else fails if ladders
- cmd parser is still iffed out. I was just tool lazy to do that
  for an RFC
- the vgpu stuff is ugly cause I didn't spend any time figuring out
  what it really wants to do
- maybe something else I'm overlooking?

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>