maintainer-tools: document the patch and merge flows and timelines
Add drm-intel patch flow (graphviz) and timeline (wavedrom) charts, and
tie them together in a restructured text source. Add a simple Makefile
to generate drm-intel.html for nice viewing (*).

This is a draft still, but much better than nothing, and provides a base
to build on.

(*) In a browser that supports SVG and has JavaScript enabled and has
    access to http://wavedrom.com/.

Signed-off-by: Jani Nikula <jani.nikula@intel.com>