dim: print branches containing commit-ish if no tag contains it
dim tc is useful for checking when and where a commit has landed, so one
can decide where, for example, a fix to that commit should be queued.

If the commit is not in a tagged upstream Linux release, fall back to
printing the i915 upstream development branches that contain it.

v2: check for remote branches (Daniel)

Signed-off-by: Jani Nikula <jani.nikula@intel.com>