dim: append commit message tags in git, not in files
Email is hard. Let that sink in.

The message piped to dim may have, among other things, base64 encoding,
and using sed to modify the commit message directly on the input falls
apart. This is also true for messages sent using git-send-email. Let
'git am' handle the hard part, and modify the commit message after the
patch has been applied.

The same holds for running checkpatch on the email message; it's
doomed. We'll leave that for later.

Signed-off-by: Jani Nikula <jani.nikula@intel.com>