drm/i915: Add some gen notes to pixel format defines
Add comments to the primary plane pixel format defines to explain which
platforms support which formats, and on which planes. Maybe next time
I won't have to dig up the spec to think about these things.

And while at it, toss a few comments to the g4x-snb video sprite format
defines too.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>